Breakout MaskDrawable
